Fixed importing http_server after renaming
